requirements.txt
PyQt5
google-generativeai
groq
chromadb
Pillow
requests
# tavily-python (if using search)
